Displaying Graphql data in frontend
Displaying the data text before displaying it with D3
