+++ package.json
@@ -1,16 +1,16 @@
 {
   "name": "power-d3",
-  "version": "0.1.2",
+  "version": "0.1.3",
   "private": true,
   "dependencies": {
-    "@apollo/client": "^3.2.3",
+    "@apollo/client": "^3.2.4",
     "@testing-library/jest-dom": "^5.11.4",
     "@testing-library/react": "^11.0.4",
     "@testing-library/user-event": "^12.1.7",
     "d3": "^6.2.0",
     "graphql": "^15.3.0",
     "react": "^16.13.1",
-    "react-admin": "^3.9.2",
+    "react-admin": "^3.9.3",
     "react-dom": "^16.13.1",
     "react-scripts": "3.4.3"
   },
