+++ src/apollo/apollo.js
@@ -1,0 +1,25 @@
+import { ApolloClient, InMemoryCache, createHttpLink } from '@apollo/client';
+
+import { setContext } from '@apollo/client/link/context';
+
+const authLink = setContext((_, { headers }) => {
+  const token = `${process.env.REACT_APP_TOKEN}`;
+
+  return {
+    headers: {
+      ...headers,
+      authorization: token ? `Bearer ${token}` : '',
+    },
+  };
+});
+
+const httpLink = createHttpLink({
+  uri: `${process.env.REACT_APP_GRAPHQL_URL}`,
+});
+
+const client = new ApolloClient({
+  link: authLink.concat(httpLink),
+  cache: new InMemoryCache(),
+});
+
+export default client;
