+++ src/index.js
@@ -4,13 +4,17 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 
-// import { ApolloProvider } from '@apollo/client';
+import { ApolloProvider } from '@apollo/client';
+
+import client from './apollo/apollo';
 
 import App from './App';
 
 ReactDOM.render(
   <React.StrictMode>
-    <App />
+    <ApolloProvider client={client}>
+      <App />
+    </ApolloProvider>
   </React.StrictMode>,
   document.getElementById('root')
 );
